+++ resources/js/lang/de/settings.js
@@ -1,10 +1,12 @@
 export default {
   title: 'Einstellungen',
-  description: 'Hier können die Einstellungen der Anwendung verwaltet werden.<br>' +
-    'Bitte wählen Sie einen der nebenstehenden Menüpunkte, um die Einstellungen anzupassen.',
+  homeButton: 'Zurück zur Übersicht',
+  overview: 'Übersicht',
+  overviewDescription: 'Hier können die Einstellungen der Anwendung verwaltet werden. Bitte wählen Sie eine der Kacheln aus, um die jeweiligen Einstellungen anzupassen.',
 
   roles: {
     title: 'Rollen',
+    tileDescription: 'Die Rollen vergeben Berechtigungen, gliedern die Benutzer und legen das max. an Räumen pro Nutzer fest.',
 
     new: 'Neue Rolle erstellen',
     view: 'Detaillierte Informationen für die Rolle {name}',
@@ -50,6 +52,7 @@ export default {
 
   users: {
     title: 'Benutzer',
+    tileDescription: 'Die Benutzer können sich an dem System anmelden und je nach Rolle unterschiedliche Funktionen nutzen.',
 
     new: 'Neuen Benutzer erstellen',
     view: 'Detaillierte Informationen für den Benutzer {firstname} {lastname}',
@@ -89,6 +92,8 @@ export default {
 
   roomTypes: {
     title: 'Raumarten',
+    tileDescription: 'Die Raumarten sortieren die Räume, geben diesen Icons zur schnelleren Wiedererkennung und bestimmen auf welchem Serverpool ein Meeting stattfindet.',
+
     icon: 'Icon',
     description: 'Beschreibung',
     short: 'Icon Text',
@@ -113,6 +118,7 @@ export default {
 
   application: {
     title: 'Anwendung',
+    tileDescription: 'Regelt systemweitere Einstellungen wie Logo, Wartungs-Banner und Seitengrößen.',
     logo: {
       title: 'Logo',
       uploadTitle: 'Logo hochladen (max. 500 KB)',
@@ -173,6 +179,7 @@ export default {
 
   servers: {
     title: 'Server',
+    tileDescription: 'Die Server stellen die BBB Infrastruktur für die Meetings bereit.',
     id: 'ID',
     description: 'Beschreibung',
     status: 'Status',