+++ resources/js/lang/en/settings.js
@@ -1,11 +1,12 @@
 export default {
   title: 'Settings',
-  description: 'Here you can manage the settings of the application.<br>' +
-    'Please select one of the menu items on the left to adjust the settings.',
+  homeButton: 'Back to the overview',
+  overview: 'Overview',
+  overviewDescription: 'Here you can manage the settings of the application. Please select one of the menu items on the left to adjust the settings.',
 
   roles: {
     title: 'Roles',
-
+    tileDescription: 'The roles assign permissions, organise the users and define the maximum number of rooms per user.',
     new: 'Create new role',
     view: 'Detailed information for the role {name}',
     edit: 'Edit role {name}',
@@ -49,7 +50,7 @@ export default {
 
   users: {
     title: 'Users',
-
+    tileDescription: 'Users can login to the system and use different features depending on their role.',
     new: 'Create new user',
     view: 'Detailed information for the user {firstname} {lastname}',
     edit: 'Edit user {firstname} {lastname}',
@@ -87,6 +88,7 @@ export default {
 
   roomTypes: {
     title: 'Room types',
+    tileDescription: 'The room types organize the rooms, give them icons for quicker recognition and determine on which server pool a meeting takes place.',
     icon: 'Icon',
     description: 'Description',
     short: 'Icon text',
@@ -111,6 +113,7 @@ export default {
 
   application: {
     title: 'Application',
+    tileDescription: 'Controls system-wide settings such as logo, maintenance banner and page sizes.',
     logo: {
       title: 'Logo',
       uploadTitle: 'Upload a logo (max. 500 KB)',
@@ -171,6 +174,7 @@ export default {
 
   servers: {
     title: 'Server',
+    tileDescription: 'The servers provide the BBB infrastructure for the meetings.',
     id: 'ID',
     description: 'Description',
     status: 'Status',