+++ resources/js/views/App.vue
@@ -34,11 +34,17 @@
         <b-collapse id="nav-collapse" is-nav>
           <b-navbar-nav>
             <b-nav-item :to="{ name: 'rooms.index' }" v-if='isAuthenticated'>{{ $t('rooms.rooms') }}</b-nav-item>
+            <can v-if='isAuthenticated' method='manage' policy='SettingPolicy'>
+              <b-nav-item :to="{ name: 'settings' }">
+                {{ $t('settings.title') }}
+              </b-nav-item>
+            </can>
           </b-navbar-nav>
 
           <!-- Right aligned nav items -->
           <b-navbar-nav class="ml-auto">
             <b-nav-item :to="{ name: 'login' }" v-if='!isAuthenticated' right>{{ $t('auth.login') }}</b-nav-item>
+
             <b-nav-item-dropdown right v-if='isAuthenticated'>
               <!-- Using 'button-content' slot -->
               <template v-if='currentUser' v-slot:button-content>
@@ -48,11 +54,6 @@
               <b-dropdown-item :to="{ name: 'profile' }">
                 {{ $t('app.profile') }}
               </b-dropdown-item>
-              <can method='manage' policy='SettingPolicy'>
-                <b-dropdown-item :to="{ name: 'settings' }">
-                  {{ $t('settings.title') }}
-                </b-dropdown-item>
-              </can>
               <b-dropdown-divider></b-dropdown-divider>
 
               <b-dropdown-item @click="logout">{{ $t('auth.logout') }}</b-dropdown-item>