+++ tests/Frontend/mixins/ActionsColumn.spec.js
@@ -37,11 +37,11 @@ describe('ActionsColumn', function () {
     PermissionService.setCurrentUser({ permissions: ['users.delete'] });
     await Vue.nextTick();
 
-    expect(view.vm.tableFields).toEqual([{ key: 'actions', label: 'app.actions', sortable: false }]);
+    expect(view.vm.tableFields).toEqual([{ key: 'actions', label: 'app.actions', sortable: false, thClass: 'actionColumn', thStyle: '', tdClass: 'actionButton' }]);
     PermissionService.setCurrentUser({ permissions: ['users.delete'] });
     await Vue.nextTick();
 
-    expect(view.vm.tableFields).toEqual([{ key: 'actions', label: 'app.actions', sortable: false }]);
+    expect(view.vm.tableFields).toEqual([{ key: 'actions', label: 'app.actions', sortable: false, thClass: 'actionColumn', thStyle: '', tdClass: 'actionButton' }]);
     PermissionService.setCurrentUser({ permissions: [] });
     await Vue.nextTick();
 
@@ -57,6 +57,47 @@ describe('ActionsColumn', function () {
     await Vue.nextTick();
 
     expect(view.vm.tableFields).toEqual([]);
+    PermissionService.setCurrentUser(oldUser);
+  });
+
+  it('changes class and style based on data', async function () {
+    const oldUser = PermissionService.currentUser;
+
+    const Test = {
+      mixins: [ActionsColumn],
+      render () {},
+      computed: {
+        tableFields () {
+          if (this.actionColumnVisible) {
+            return [this.actionColumnDefinition];
+          }
+
+          return [];
+        }
+      },
+      data () {
+        return {
+          actionColumnThClass: 'testClass',
+          actionColumnThStyle: 'testStyle',
+          actionColumnTdClass: 'testClass2',
+          actionPermissions: ['test.perm']
+        };
+      }
+    };
+
+    const view = mount(Test, {
+      mocks: {
+        $t: key => key
+      }
+    });
+
+    expect(view.vm.tableFields).toEqual([]);
+
+    PermissionService.setCurrentUser({ permissions: ['test.perm'] });
+    await Vue.nextTick();
+
+    expect(view.vm.tableFields).toEqual([{ key: 'actions', label: 'app.actions', sortable: false, thClass: 'testClass', thStyle: 'testStyle', tdClass: 'testClass2' }]);
+    await Vue.nextTick();
     PermissionService.setCurrentUser(oldUser);
   });
 });