+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added file upload for room owner and file download for participants ([#4], [#18])
 - Added agreement check before file download ([#50],[#54])
 - Added global settings, .env as default; added logo path and room limit ([#34],[#36])
-- Added setting menu for administrators ([#35], [#38])
+- Added setting menu for administrators ([#35], [#38], [#97],[#100])
 - Added room type management pages to admin settings ([#62], [#73])
 - Added a middleware to check whether the request is in sync with model of the database and not stale ([#40], [#41])
 - Added management of users and profile page with possibility to disable audio echo test ([#10], [#66], [#67], [#87])
@@ -76,6 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#88]: https://github.com/THM-Health/PILOS/pull/88
 [#89]: https://github.com/THM-Health/PILOS/issues/89
 [#94]: https://github.com/THM-Health/PILOS/pull/94
+[#97]: https://github.com/THM-Health/PILOS/issues/97
+[#100]: https://github.com/THM-Health/PILOS/pull/100
 [#101]: https://github.com/THM-Health/PILOS/issues/101
 [#102]: https://github.com/THM-Health/PILOS/pull/102
 [#103]: https://github.com/THM-Health/PILOS/issues/103